# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sonnenBatterie API",
    "description": "The sonnenBatterie API allows to remote control a sonnenBatterie.\n                           \nMost API endpoints require an authentication token.\n                           This token is available in the sonnenBatterie Dashboard (‘Software-Integration’).\n                           Unprotected endpoints are marked below.\n                           \n\nExample request:\n                           \n`curl --header 'Auth-Token: TOKEN' http://SYSTEM-IP/api/v2/latestdata`",
    "license": {
      "name": "Proprietary",
      "url": "https://www.sonnen.de"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://IP",
      "description": "sonnenBatterie"
    }
  ],
  "paths": {
...
    "/api/v2/configurations": {
      "get": {
        "tags": [
          "Read API"
        ],
        "summary": "Get all configurations",
        "description": "Get all configurations\n\nGets the values of all allowed configurations. The allowed configurations are:\n- CM_MarketingModuleCapacity\n- CN_CascadingRole\n- DE_Software\n- EM_OperatingMode\n- EM_Prognosis_Charging\n- EM_RE_ENABLE_MICROGRID\n- EM_ToU_Schedule\n- EM_USER_INPUT_TIME_ONE\n- EM_USER_INPUT_TIME_THREE\n- EM_USER_INPUT_TIME_TWO\n- EM_USOC\n- EM_US_CHP_Max_SOC\n- EM_US_CHP_Min_SOC\n- EM_US_GENRATOR_TYPE\n- EM_US_GEN_POWER_SET_POINT\n- EM_US_RE_ENABLE_MICROGRID\n- EM_US_USER_INPUT_TIME_ONE\n- EM_US_USER_INPUT_TIME_THREE\n- EM_US_USER_INPUT_TIME_TWO\n- IC_BatteryModules\n- IC_InverterMaxPower_w\n- NVM_PfcFixedCosPhi\n- NVM_PfcIsFixedCosPhiActive\n- NVM_PfcIsFixedCosPhiLagging\n- SH_HeaterOperatingMode\n- SH_HeaterTemperatureMax\n- SH_HeaterTemperatureMin",
        "operationId": "GetConfigurations",
        "responses": {
          "200": {
            "description": "Get all Configurations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configurations"
                },
                "example": {
                  "CM_MarketingModuleCapacity": "5500",
                  "CN_CascadingRole": "secondary",
                  "DE_Software": "1.8.3",
                  "EM_CHP_Max_SOC": "90",
                  "EM_CHP_Min_SOC": null,
                  "EM_OperatingMode": "1",
                  "EM_Prognosis_Charging": "0",
                  "EM_RE_ENABLE_MICROGRID": "0",
                  "EM_ToU_Schedule": "[{\"grid\":\"1\",\"start\":\"07:00\",\"stop\":\"08:00\",\"charge\":\"07:31\"}]",
                  "EM_USER_INPUT_TIME_ONE": "00:00",
                  "EM_USER_INPUT_TIME_TWO": "00:00",
                  "EM_USOC": "0",
                  "EM_US_GENRATOR_TYPE": "automatic",
                  "EM_US_GEN_POWER_SET_POINT": "0",
                  "IC_BatteryModules": "4",
                  "IC_InverterMaxPower_w": "3300",
                  "NVM_PfcFixedCosPhi": "1",
                  "NVM_PfcIsFixedCosPhiActive": "0",
                  "NVM_PfcIsFixedCosPhiLagging": "1",
                  "SH_HeaterOperatingMode": "0",
                  "SH_HeaterTemperatureMax": "65",
                  "SH_HeaterTemperatureMin": "42"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/PublicUnauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "Write API"
        ],
        "summary": "Update configurations",
        "description": "Update configurations\n\nSets the value of one or more configurations. The writable configurations are:\n- EM_OperatingMode\n- EM_Prognosis_Charging\n- EM_RE_ENABLE_MICROGRID\n- EM_ToU_Schedule\n- EM_USER_INPUT_TIME_ONE\n- EM_USER_INPUT_TIME_THREE\n- EM_USER_INPUT_TIME_TWO\n- EM_USOC\n- EM_US_CHP_Max_SOC\n- EM_US_CHP_Min_SOC\n- EM_US_GENRATOR_TYPE\n- EM_US_GEN_POWER_SET_POINT\n- EM_US_RE_ENABLE_MICROGRID\n- EM_US_USER_INPUT_TIME_ONE\n- EM_US_USER_INPUT_TIME_THREE\n- EM_US_USER_INPUT_TIME_TWO\n- NVM_PfcFixedCosPhi\n- NVM_PfcIsFixedCosPhiActive\n- NVM_PfcIsFixedCosPhiLagging\n- SH_HeaterOperatingMode\n- SH_HeaterTemperatureMax\n- SH_HeaterTemperatureMin\n\nThe operating mode (EM_OperatingMode) can not be changed if VPP is active.\n\nPossible values for EM_USOC are:\n* 0 (disable backup buffer)\n* 5 - 100 (backup buffer percentage)\n\n**Examples:**\n\n`curl -X PUT -d NVM_PfcFixedCosPhi=1 --header 'Auth-Token: TOKEN' http://SYSTEM-IP/api/v2/configurations`\n\n`curl -X PUT -d NVM_PfcFixedCosPhi=1 -d NVM_PfcIsFixedCosPhiActive=0 --header 'Auth-Token: TOKEN' http://SYSTEM-IP/api/v2/configurations`\n",
        "operationId": "PutConfiguration",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Configurations"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResponseBody"
                },
                "example": {
                  "EM_OperatingMode": "2"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/PublicUnauthorized"
          },
          "403": {
            "description": "403 Forbidden"
          }
        }
      }
    },
...
    "/api/v2/status": {
      "get": {
        "tags": [
          "Read API"
        ],
        "summary": "Get status data",
        "description": "Get status data\n\nThis endpoint does not require an authentication token.\n\n| Variable | Example Value | Unit | Description | Data type |\n| --- | --- | --- | --- | --- |\n| Apparent\\_output | 226 | VA, volt-ampere | All AC output of apparent power in VA | Integer |\n| BackupBuffer | 0 | %, percentage | Backup-buffer in percentage that is set on the system. | Integer |\n| BatteryCharging | false | \\- | Boolean that indicates the charge status. True if charging | Boolean |\n| BatteryDischarging | false | \\- | Boolean that indicates the discharge status. True if discharging | Boolean |\n| Consumption\\_Avg | 223 | W, watt | House consumption in watts, average over the last 60s | Integer |\n| Consumption\\_W | 232 | W, watt | House consumption in watts, direct measurement | Integer |\n| Fac | 49.999000549316400 | Hz, hertz | AC frequency in hertz | Float |\n| FlowConsumptionBattery | false | \\- | Boolean that indicates the energy flow at the installation site. True if battery feeds the consumption | Boolean |\n| FlowConsumptionGrid | true | \\- | Boolean that indicates the energy flow at the installation site. True if grid feeds the consumption | Boolean |\n| FlowConsumptionProduction | true | \\- | Boolean that indicates the energy flow at the installation site. True if production feeds the consumption | Boolean |\n| FlowGridBattery | false | \\- | Boolean that indicates the energy flow at the installation site. True if battery is charging from grid | Boolean |\n| FlowProductionBattery | false | \\- | Boolean that indicates the energy flow at the installation site. True if production is charging the battery | Boolean |\n| FlowProductionGrid | false | \\- | Boolean that indicates the energy flow at the installation site. True if production feeds into the grid | Boolean |\n| GridFeedIn\\_W | \\-208 | W, watt | Grid Feed in negative is consumption and positive is feed in | Integer |\n| IsSystemInstalled | 1 | \\- | System is installed or not | Integer |\n| OperatingMode | 2 | \\- | Operating mode that is set on the system: 1: Manual charging or discharging via API; 2: Automatic Self Consumption. Default.| Integer |\n| Pac\\_total\\_W | \\-5 | W, watt | AC Power greater than ZERO is discharging Inverter AC Power less than ZERO is charging | Signed Integer |\n| Production\\_W | 28 | W, watt | PV production in watts | Integer |\n| RSOC | 4 | %, percentage | Relative state of charge | Integer |\n| RemainingCapacity\\_W | 674 | Wh | Remaining capacity based on RSOC | Integer |\n| Sac1 | 75 | VA, volt-ampere | Output of apparent power in VA on Phase 1 | Integer |\n| Sac2 | 75 | VA, volt-ampere | Output of apparent power in VA on Phase 2 | Integer |\n| Sac3 | 76 | VA, volt-ampere | Output of apparent power in VA on Phase 3 | Integer |\n| SystemStatus | OnGrid | \\- | String that indicates if the system is connected to the grid (“OnGrid”) or disconnected (“OffGrid”) | String |\n| Timestamp | 2020-12-10 11:26:01 | \\- | Local system time | String |\n| USOC | 0 | %, percentage | User state of charge | Integer |\n| Uac | 230 | V, volt | AC voltage in volts | Integer |\n| Ubat | 50 | V, volt | Battery voltage in volts | Integer |\n| dischargeNotAllowed | false | \\- | Boolean that indicates the discharge status. True if no discharge allowed, based on battery maintenance | Boolean |\n| generator\\_autostart | false | \\- | Boolean that indicates the autostart setting of the generator. | Boolean |\n",
        "operationId": "Status",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponseBody"
                },
                "example": {
                  "Apparent_output": 96,
                  "BackupBuffer": "0",
                  "BatteryCharging": false,
                  "BatteryDischarging": false,
                  "Consumption_W": 0,
                  "Fac": 50.0239143371582,
                  "FlowConsumptionBattery": false,
                  "FlowConsumptionGrid": false,
                  "FlowConsumptionProduction": false,
                  "FlowGridBattery": false,
                  "FlowProductionBattery": false,
                  "FlowProductionGrid": false,
                  "GridFeedIn_W": -3,
                  "IsSystemInstalled": 1,
                  "OperatingMode": "2",
                  "Pac_total_W": 7,
                  "Production_W": 0,
                  "RSOC": 21,
                  "Sac1": 96,
                  "Sac2": null,
                  "Sac3": null,
                  "SystemStatus": "OnGrid",
                  "Timestamp": "2020-03-26 17:10:06",
                  "USOC": 12,
                  "Uac": 238,
                  "Ubat": 209,
                  "dischargeNotAllowed": false,
                  "generator_autostart": false
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/PublicUnauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Configurations": {
        "type": "object",
        "additionalProperties": {
          "type": "string",
          "description": "A map of configuration entries with key and value.",
          "nullable": true
        }
      },
...
      "GetResponseBody": {
        "type": "object",
        "additionalProperties": {
          "type": "string",
          "nullable": true
        }
      },
      "StatusResponseBody": {
        "type": "object",
        "properties": {
          "Apparent_output": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "BackupBuffer": {
            "type": "string",
            "nullable": true
          },
          "BatteryCharging": {
            "type": "boolean",
            "nullable": true
          },
          "BatteryDischarging": {
            "type": "boolean",
            "nullable": true
          },
          "Consumption_Avg": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Consumption_W": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Fac": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "FlowConsumptionBattery": {
            "type": "boolean",
            "nullable": true
          },
          "FlowConsumptionGrid": {
            "type": "boolean",
            "nullable": true
          },
          "FlowConsumptionProduction": {
            "type": "boolean",
            "nullable": true
          },
          "FlowGridBattery": {
            "type": "boolean",
            "nullable": true
          },
          "FlowProductionBattery": {
            "type": "boolean",
            "nullable": true
          },
          "FlowProductionGrid": {
            "type": "boolean",
            "nullable": true
          },
          "GridFeedIn_W": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "IsSystemInstalled": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "OperatingMode": {
            "type": "string",
            "nullable": true
          },
          "Pac_total_W": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Production_W": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "RSOC": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "RemainingCapacity_Wh": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Sac1": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Sac2": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Sac3": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "SystemStatus": {
            "type": "string",
            "nullable": true
          },
          "Timestamp": {
            "type": "string",
            "nullable": true
          },
          "USOC": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Uac": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "Ubat": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "dischargeNotAllowed": {
            "type": "boolean",
            "nullable": true
          },
          "generator_autostart": {
            "type": "boolean",
            "nullable": true
          }
        }
      }
    },
    "responses": {
...
      "InternalServerError": {
        "description": "Internal server error has occurred.\nPlease read the details in the response body.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "description": "Internal server error has occurred.\nPlease read the details in the response body.",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "internal server error"
                }
              }
            }
          }
        }
      },
      "PublicUnauthorized": {
        "description": "Request is unauthorized. The authentication token is invalid or missing.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "description": "Request is unauthorized. The authentication token is invalid or missing.",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "API-Token": {
        "type": "apiKey",
        "in": "header",
        "name": "Auth-Token"
      }
    }
  },
  "security": [
    {
      "API-Token": []
    }
  ],
...
}